requirements-dev.txt
sphinx==8.0.0
black==25.11.0
flake8==7.3.0
isort==5.12.0
mypy==1.8.0
pylint==3.2.6
setuptools==80.9.0
pytest>=8.2
pytest-asyncio>=0.23
python-dotenv>=1.0
